+++ server/src/main/scala/org/scassandra/server/priming/routes/PrimingQueryRoute.scala
@@ -15,6 +15,7 @@
  */
 package org.scassandra.server.priming.routes
 
+import org.scassandra.server.priming.json.PrimingJsonImplicits
 import spray.routing.HttpService
 import com.typesafe.scalalogging.LazyLogging
 import spray.http.StatusCodes
