+++ server/src/test/scala/org/scassandra/server/e2e/verification/PreparedStatementExecutionVerificationTest.scala
@@ -21,11 +21,12 @@ import org.scalatest.concurrent.ScalaFutures
 import org.scassandra.server.AbstractIntegrationTest
 import org.scassandra.server.cqlmessages.ONE
 import org.scassandra.server.priming.PreparedStatementExecution
+import org.scassandra.server.priming.json.PrimingJsonImplicits
 import spray.json._
 
 class PreparedStatementExecutionVerificationTest extends AbstractIntegrationTest with ScalaFutures {
 
-  import org.scassandra.server.priming.PrimingJsonImplicits._
+  import PrimingJsonImplicits._
 
   val primePreparedSinglePath = "http://localhost:8043/prepared-statement-execution"
